Mechanical changes in core/network: - Move GenerateVirtualMACAddress out of the address module and into   network. - Rename SubnetSet to IDSet, as it is a generic for network.ID and move   its tests into the network suite.
